+++ src/main/java/ru/softdarom/qrcheck/events/model/dto/TicketDto.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.Generated;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Data
 @Generated
@@ -24,9 +25,9 @@ public class TicketDto {
 
     @NotNull
     @JsonProperty("cost")
-    private Double cost;
+    private BigDecimal cost;
 
     @JsonProperty("price")
-    private Double price;
+    private BigDecimal price;
 
 }
